# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Real-Time Analytics Pipeline API",
    "version": "1.0.0",
    "description": "Enterprise-grade real-time analytics pipeline API. Supports event ingestion, analytics queries, dashboard management, and real-time WebSocket streaming.",
    "contact": {
      "name": "Engineering Team",
      "email": "engineering@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://api.example.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "VALIDATION_ERROR" },
              "message": { "type": "string", "example": "Validation failed" },
              "details": { "type": "object", "example": {} }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "EventCreateRequest": {
        "type": "object",
        "required": ["event_type", "source"],
        "properties": {
          "event_type": { "type": "string", "enum": ["page_view", "click", "signup", "purchase", "login", "logout", "error", "custom"], "description": "Type of event" },
          "source": { "type": "string", "minLength": 1, "maxLength": 255, "example": "web-app" },
          "payload": { "type": "object", "example": {"url": "/home", "duration_ms": 4500} },
          "user_id": { "type": "string", "nullable": true },
          "session_id": { "type": "string", "nullable": true },
          "ip_address": { "type": "string", "nullable": true },
          "user_agent": { "type": "string", "nullable": true },
          "timestamp": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "EventResponse": {
        "type": "object",
        "properties": {
          "event_id": { "type": "string", "format": "uuid" },
          "event_type": { "type": "string" },
          "source": { "type": "string" },
          "payload": { "type": "object" },
          "timestamp": { "type": "string", "format": "date-time" },
          "user_id": { "type": "string", "nullable": true },
          "session_id": { "type": "string", "nullable": true },
          "status": { "type": "string" }
        }
      },
      "EventIngestResponse": {
        "type": "object",
        "properties": {
          "event_id": { "type": "string", "format": "uuid" },
          "message_id": { "type": "string" },
          "status": { "type": "string", "example": "accepted" }
        }
      },
      "BatchIngestResponse": {
        "type": "object",
        "properties": {
          "count": { "type": "integer" },
          "event_ids": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "string", "example": "accepted" }
        }
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array" },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total_pages": { "type": "integer" },
          "has_next": { "type": "boolean" },
          "has_previous": { "type": "boolean" }
        }
      },
      "EventListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/PaginatedResponse" },
          {
            "type": "object",
            "properties": {
              "items": { "type": "array", "items": { "$ref": "#/components/schemas/EventResponse" } }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/v1/events/ingest": {
      "post": {
        "operationId": "ingest_event",
        "summary": "Ingest a single event",
        "description": "Ingests a single analytics event into the pipeline. The event is queued for async processing.",
        "tags": ["Events"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EventCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event accepted for processing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventIngestResponse" }
              }
            }
          },
          "422": {
            "description": "Invalid event data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/ingest/batch": {
      "post": {
        "operationId": "ingest_events_batch",
        "summary": "Batch ingest events",
        "description": "Ingests multiple analytics events in a single request for higher throughput.",
        "tags": ["Events"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/EventCreateRequest" },
                "minItems": 1,
                "maxItems": 500
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Events accepted for processing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BatchIngestResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/events": {
      "get": {
        "operationId": "list_events",
        "summary": "List events",
        "description": "Retrieves analytics events with filtering and pagination.",
        "tags": ["Events"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "start", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" }, "description": "Start of time range" },
          { "name": "end", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" }, "description": "End of time range" },
          { "name": "event_type", "in": "query", "schema": { "type": "string" } },
          { "name": "source", "in": "query", "schema": { "type": "string" } },
          { "name": "user_id", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventListResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}": {
      "get": {
        "operationId": "get_event",
        "summary": "Get event by ID",
        "description": "Returns a single analytics event by its unique ID.",
        "tags": ["Events"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "event_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Event found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventResponse" }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}